+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## 8.1.0-beta
 
+- Make `Http.AppendQueryToUrl` public (closes #1325)
 - Add `PreferOptionals` parameter to `JsonProvider` and `XmlProvider` (defaults to `true` to match existing behavior; set to `false` to use empty string or `NaN` for missing values, like the CsvProvider default) (closes #649)
 
 ## 8.0.0 - Feb 25 2026